modify filters and add some fixes (m -> mm)

+++ services/layout/tests/rules/household/fifth_level_test.go
@@ -10,7 +10,7 @@ func TestHouseholdLevel5(t *testing.T) {
 	_, robotFilter := placeRobotVacuumOnLevel(t, "5")
 
 	assert.Equal(t, float64(65), robotFilter.NoiseLevelDB)
-	assert.Equal(t, float64(11000), robotFilter.SuctionPowerPA)
+	assert.Equal(t, 11000, robotFilter.SuctionPowerPA)
 	assert.Equal(t, "lidar_ai", robotFilter.NavigationType)
 	assert.True(t, robotFilter.RoomMapping)
 	assert.True(t, robotFilter.WetCleaning)
